+++ src/main/java/com/example/oatnote/_commons/util/JwtUtil.java
@@ -1,4 +1,4 @@
-package com.example.oatnote.util;
+package com.example.oatnote._commons.util;
 
 import java.nio.charset.StandardCharsets;
 import java.security.Key;
