+++ src/main/java/com/example/oatnote/domain/memotag/rabbitmq/MemoTagMessageProducer.java
@@ -9,7 +9,7 @@ import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
-import com.example.oatnote.domain.memotag.rabbitmq.dto.MemoTagMessage;
+import com.example.oatnote._commons.message.MemoTagMessage;
 import com.example.oatnote.domain.memotag.service.client.dto.AICreateTagsResponse;
 import com.example.oatnote.domain.memotag.service.memo.model.Memo;
 import com.example.oatnote.web.exception.server.OatExternalServiceException;
