+++ src/main/java/com/example/oatnote/domain/user/service/UserService.java
@@ -22,7 +22,7 @@ import com.example.oatnote.domain.user.service.email.EmailVerificationService;
 import com.example.oatnote.domain.user.service.model.User;
 import com.example.oatnote.event.RegisterUserEvent;
 import com.example.oatnote.event.WithdrawUserEvent;
-import com.example.oatnote.util.JwtUtil;
+import com.example.oatnote._commons.util.JwtUtil;
 import com.example.oatnote.web.exception.auth.OatInvalidPasswordException;
 import com.example.oatnote.web.exception.client.OatDataNotFoundException;
 import com.example.oatnote.web.exception.client.OatIllegalArgumentException;
